requirements.txt
flask
pypdf
numpy
python-dotenv
google-genai
google-generativeai